// app/modules/reporting/models/api/parameter-enums/report-category-parameter.ts
enum ReportCategoryParameter {
  TEST = 'test',
  ANALYTICS = 'analytics',
  FLOW = 'flow',
  QOS = 'qos',
  IPSLA = 'IPSLA',
  VIPTELA =  'viptela',
  MEDIANET = 'medianet',
  AVC = 'avc',
  TOPOLOGY = 'topology',
  ALERT = 'alert',
  STATUS_HEALTH = 'status',
}

export default ReportCategoryParameter;